+++ src/utils/getStateFromCoordinates.test.ts
@@ -1,75 +1,24 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-import type { FeatureCollection } from 'geojson';
-
-const mockGeoJSON: FeatureCollection = {
-    type: 'FeatureCollection',
-    features: [
-        {
-            type: 'Feature',
-            properties: { name: 'MockState' },
-            geometry: {
-                type: 'Polygon',
-                coordinates: [
-                    [
-                        [-105.1, 39.9],
-                        [-104.9, 39.9],
-                        [-104.9, 40.1],
-                        [-105.1, 40.1],
-                        [-105.1, 39.9],
-                    ],
-                ],
-            },
-        },
-    ],
-};
-
-vi.mock('fs', async () => {
-    const actual = await vi.importActual<typeof import('fs')>('fs');
-    return {
-        ...actual,
-        readFileSync: vi.fn(() => JSON.stringify(mockGeoJSON)),
-    };
-});
-
-vi.mock('path', async () => {
-    const actual = await vi.importActual<typeof import('path')>('path');
-    return {
-        ...actual,
-        join: vi.fn(() => 'mock/path/us-states.geojson'),
-    };
-});
-
-vi.resetModules();
-
+import { describe, it, expect } from 'vitest';
 import { getStateFromCoordinates } from './getStateFromCoordinates';
 
+const testCases = [
+  { lat: 39.7392, lng: -104.9903, expected: 'Colorado' }, 
+  { lat: 34.0522, lng: -118.2437, expected: 'California' }, 
+  { lat: 40.7128, lng: -74.006, expected: 'New York' },
+  { lat: 41.8781, lng: -87.6298, expected: 'Illinois' }, 
+  { lat: 47.6062, lng: -122.3321, expected: 'Washington' }, 
+];
+
 describe('getStateFromCoordinates', () => {
-  it('returns the correct state name when point is inside polygon', async () => {
-    vi.resetModules();
-    const { getStateFromCoordinates, __clearStateCache } = await import('./getStateFromCoordinates');
-    __clearStateCache();
-    const result = getStateFromCoordinates(40.0, -105.0);
-    expect(result).toBe('MockState');
+  it('returns the correct state for known coordinates', () => {
+    for (const { lat, lng, expected } of testCases) {
+      const result = getStateFromCoordinates(lat, lng);
+      expect(result).toBe(expected);
+    }
   });
 
-  it('returns null if feature has no name', async () => {
-    mockGeoJSON.features[0].properties = {};
-    vi.resetModules();
-    const { getStateFromCoordinates, __clearStateCache } = await import('./getStateFromCoordinates');
-    __clearStateCache();
-    const result = getStateFromCoordinates(40.0, -105.0);
+  it('returns null for out-of-bounds coordinates', () => {
+    const result = getStateFromCoordinates(0, 0); // Not in US
     expect(result).toBeNull();
-    mockGeoJSON.features[0].properties = { name: 'MockState' };
-  });
-
-  it('returns null if feature has no name', async () => {
-    mockGeoJSON.features[0].properties = {};
-    vi.resetModules();
-    const { getStateFromCoordinates, __clearStateCache } = await import('./getStateFromCoordinates');
-    __clearStateCache();
-    const result = getStateFromCoordinates(40.0, -105.0);
-    expect(result).toBeNull();
-
-    mockGeoJSON.features[0].properties = { name: 'MockState' };
   });
 });
